% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_thin.R
\name{distance_thin}
\alias{distance_thin}
\title{Filter a vcf file based on distance between SNPs on a given scaffold}
\usage{
distance_thin(vcfR, min.distance = NULL)
}
\arguments{
\item{vcfR}{a vcfR object}
}
\value{
An identical vcfR object, except that SNPs separated by less than the
specified distance have been removed from the file
}
\description{
This function requires a vcfR object as input, and returns a vcfR object filtered
to retain only SNPs greater than a specified distance apart on each scaffold.
This type of filtering is often employed to reduce linkage amongst SNPs.
}
\examples{
distance_thin(vcfR = SNPfiltR:::vcfR.example, min.distance = 1000)
}
