% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_biallelic.R
\name{filter_biallelic}
\alias{filter_biallelic}
\title{Remove SNPs with more than two alleles}
\usage{
filter_biallelic(vcfR)
}
\arguments{
\item{vcfR}{a vcfR object}
}
\value{
a vcfR object with SNPs containing more than two alleles removed
}
\description{
This function simply removes any SNPs from the vcf file for which more than
two alleles are found within your samples. Many downstream applications require
SNPs to be biallelic, so this filter is generally a good idea.
}
\examples{
filter_biallelic(vcfR=system.file("extdata","unfiltered.vcf.gz",package="SNPfiltR",mustWork=TRUE))
}
