% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hard.filter.R
\name{hard.filter.vcf}
\alias{hard.filter.vcf}
\title{Hard filter a vcf file by depth and genotype quality (gq)}
\usage{
hard.filter.vcf(vcfR, depth = NULL, gq = NULL)
}
\arguments{
\item{vcfR}{a vcfR object}

\item{depth}{an integer representing the minimum depth for genotype calls that you wish to retain
(e.g. 'depth = 5' would remove all genotypes with a sequencing depth of 4 reads or less)}

\item{gq}{an integer representing the minimum genotype quality for genotype calls that you wish to retain
#' (e.g. 'gq = 30' would remove all genotypes with a quality score of 29 or lower)}
}
\value{
The vcfR object input, with the sites failing specified filters converted to 'NA'
}
\description{
This function requires a vcfR object as input.
There are slots for varying the n parameter across M-1, M, and M-1 (as recommended by Paris et al. 2017).
After running stacks with each of the n options, plug the output vcf files into this
function to visualize the effect of varying m on number of SNPs and loci built to
recognize which value optimizes the n parameter for your dataset at the 'R80' cutoff (Paris et al. 2017).
}
\examples{
hard.filter.vcf(vcfR=system.file("extdata","unfiltered.vcf.gz",package="SNPfiltR",mustWork=TRUE))
}
