% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_mac.R
\name{min_mac}
\alias{min_mac}
\title{Vizualise how Minor Allele Count (MAC) affects population structure, filter for MAC}
\usage{
min_mac(vcfR, popmap = NULL, min.mac = NULL)
}
\arguments{
\item{vcfR}{a vcfR object}

\item{min.mac}{an integer specifying the minimum minor allele count for a SNP to be retained (e.g. 'min.mac=3'
would remove all SNPs with a MAC of 2 or less)}
}
\value{
if 'min.mac' is not specified, will print out DAPC results. If 'min.mac' is specified, SNPs
falling below the MAC cutoff will be removed, and the filtered vcfR object will be returned.
}
\description{
This function can be run in two ways: 1) Without 'min.mac' specified. This will run DAPC for a
minimum MAC of 1,2,3,4,5,10, visualize the results, and tell you how the DAPC clustering
compares to your a priori population assignments, and how many SNPs are retained at each MAC cutoff.
2) With 'min.mac' specified. This will print your folded Site Frequency Spectrum (SFS) and
show you where your specified min. MAC count falls. It will then return your vcfR object with SNPs
falling below your min. MAC threshold removed.
MAF cutoffs can be helpful in removing spurious and uninformative loci from the dataset, but also
have the potential to bias downstream inferences. Linck and Battey (2019) have an excellent paper on
just this topic. From the paper- "We recommend researchers using model‐based programs to describe
population structure observe the following best practices: (a) duplicate analyses with nonparametric
methods suchas PCA and DAPC with cross validation (b) exclude singletons (c) compare alignments with
multiple assembly parameters When seeking to exclude only singletons in alignments with missing data
(a ubiquitous problem for reduced‐representation library preparation methods), it is preferable to filter
by the count (rather than frequency) of the minor allele, because variation in the amount of missing data
across an alignment will cause a static frequency cutoff to remove different SFS classes at different sites".
Based on the differences between DAPC runs and the variation in number of SNPs retained, you should
be able to make an informed decision about whether to implement a MAC cutoff for your dataset.
Note: previous filtering steps may have converted called genotypes to 'NAs' and resulting in invariant SNPs
(MAC =0) for this reason it's a good idea to at least run min.mac(vcfR, min.mac=1).
}
\examples{
min_mac(vcfR=system.file("extdata","unfiltered.vcf.gz",package="SNPfiltR",mustWork=TRUE), min.mac=2)
}
