% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{popmap}
\alias{popmap}
\title{Popmap for example scrub-jay vcfR file}
\format{
A data frame with 20 rows and 2 variables
}
\usage{
data(popmap)
}
\description{
A dataset containing the sample name and population assignment for the 20
scrub-jay samples in SNPfilR::vcfR.example . The variables are as follows:
}
\details{
\itemize{
\item id. unique sample identifier
\item pop. population assignment for each individual
}
}
\keyword{datasets}
