% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{vcfR.example}
\alias{vcfR.example}
\title{Example scrub-jay vcfR file}
\format{
A vcfR object containing 500 SNPs for 20 individuals
}
\usage{
data(vcfR.example)
}
\description{
A vcfR object containing 500 SNPs for 20 individuals. Species assignments for
each individual can be accessed via SNPfiltR::popmap
}
\keyword{datasets}
